fix: collapse get :id with  get all clients

+++ internal/middleware/client.go
@@ -52,12 +52,12 @@ func (cm ClientMiddleware) ClientJWT(c *fiber.Ctx) error {
 		return unauthorized
 	}
 
-	_, err = cm.ClientService.Get(cdto.GetClientReq{Id: int64(clientVerified.ID)})
+	_, err = cm.ClientService.Get(cdto.GetClientReq{Id: clientVerified.ID})
 	if err != nil {
 		return unauthorized
 	}
 
-	c.Locals("clientId", int64(clientVerified.ID))
+	c.Locals("clientId", clientVerified.ID)
 
 	return c.Next()
 }
